Added changes to .yml to support blue green / feature testing. updated testing.py
